+++ components/core/ui/ktx/src/main/java/com/flipperdevices/core/ui/ktx/ComposableKeyType.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.core.ui.ktx
 
-import androidx.annotation.ColorRes
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.defaultMinSize
@@ -13,33 +12,33 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.flipperdevices.bridge.dao.api.R as DaoR
+import com.flipperdevices.bridge.dao.api.R
 import com.flipperdevices.bridge.dao.api.model.FlipperFileType
+import com.flipperdevices.bridge.dao.api.model.FlipperFileType.Companion.colorByFlipperFileType
 import com.flipperdevices.core.ui.res.R as DesignSystem
+import com.flipperdevices.core.ui.theme.LocalPallet
+import com.flipperdevices.core.ui.theme.LocalTypography
 
 @Composable
 @Preview
 fun ComposableKeyType(
     type: FlipperFileType? = FlipperFileType.NFC,
-    @ColorRes colorId: Int? = type?.color,
+    colorKey: Color = colorByFlipperFileType(type),
     minWidth: Dp? = 110.dp
 ) {
-    val icon = type?.icon ?: DaoR.drawable.ic_fileformat_unknown
-    val color = colorId ?: DaoR.color.fileformat_color_unknown
+    val icon = type?.icon ?: DesignSystem.drawable.ic_fileformat_unknown
     val title = type?.humanReadableName
-        ?: stringResource(DaoR.string.fileformat_unknown)
+        ?: stringResource(R.string.fileformat_unknown)
 
     var rowModifier = Modifier
         .clip(RoundedCornerShape(bottomEnd = 18.dp))
-        .background(colorResource(color))
+        .background(colorKey)
 
     if (minWidth != null) {
         rowModifier = rowModifier.defaultMinSize(minWidth = minWidth)
@@ -58,13 +57,10 @@ fun ComposableKeyType(
         )
 
         Text(
-            modifier = Modifier.padding(
-                end = 14.dp
-            ),
+            modifier = Modifier.padding(end = 14.dp),
             text = title,
-            fontSize = 14.sp,
-            fontWeight = FontWeight.W500,
-            color = colorResource(DesignSystem.color.black_100)
+            color = LocalPallet.current.text100,
+            style = LocalTypography.current.bodyM14
         )
     }
 }